requirements.txt
Pillow
bleak
gpiozero
lgpio
spidev

